+++ amplify/addcontent.ts
@@ -1,4 +1,6 @@
 "use server";
+import { DynamoDBClient, ScanCommand } from '@aws-sdk/client-dynamodb';
+import { DynamoDBDocumentClient } from '@aws-sdk/lib-dynamodb';
 import { S3Client, PutObjectCommand } from '@aws-sdk/client-s3';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 import 'dotenv/config';
@@ -10,6 +12,8 @@ const credentials = {
         secretAccessKey: process.env.SECRET_ACCESS_KEY!,
       };
 const s3 = new S3Client({ region, credentials });
+
+const dynamoClient = DynamoDBDocumentClient.from(new DynamoDBClient([{ region, credentials, }]));
 
 export const  addURL = async ( filename: string, contentType: string) => {   
   try {
@@ -28,3 +32,14 @@ export const  addURL = async ( filename: string, contentType: string) => {
   }
 }
 
+export const viewAllItems = async () => {
+  const tableName="moderation-results"
+  const command = new ScanCommand({ TableName: tableName });
+  const response = await dynamoClient.send(command);
+  const res = response.Items; 
+  res?.map((e) => {
+    console.log(e.moderation_results+"=>"+e )    
+  })
+  //console.log("Items:", phantu.confidence_score);
+  return response.Items;
+}
